requirements.txt
apache-airflow==2.10.3
apache-airflow-providers-postgres==5.10.0
requests==2.32.3
psycopg2-binary==2.9.9

